+++ docs/history.rst
@@ -13,6 +13,7 @@ Version History
 - Fail gracefully when a transcoder raises a :exc:`TypeError` or :exc:`ValueError` when encoding
   the response
 - Remove setuptools, replace with :pep:`517` and hatch_
+- Replace flake8 with ruff
 
 .. _application/x-www-formurlencoded: https://url.spec.whatwg.org/#application/x-www-form-urlencoded
 .. _hatch: https://hatch.pypa.io/latest/